File Restructure 2.
Moved Upgrade Finder files into Panels (panels, item cards)

Moved LogImport Modules to their own folder outside of Cooldown Planner.

Standardised the updated Duration function across modules.
