+++ src/Modules/UpgradeFinder/Panels/PanelWorldBosses.js
@@ -5,8 +5,8 @@ import ItemUpgradeCard from "./ItemUpgradeCard";
 import UpgradeFinderBossImages from "./BossImages";
 import "./Panels.css";
 import { useTranslation } from "react-i18next";
-import { encounterDB } from "../Player/InstanceDB";
-import { filterItemListBySource, getDifferentialByID } from "../Engine/ItemUtilities";
+import { encounterDB } from "../../Player/InstanceDB";
+import { filterItemListBySource, getDifferentialByID } from "../../Engine/ItemUtilities";
 
 const useStyles = makeStyles((theme) => ({
   root: {
