+++ src/Modules/UpgradeFinder/Panels/Panels.css
